Add command to synchronize library
The library:sync command allows to synchronize the existing files in a
local library with the favorites on the OC ReMix station on Rainwave.cc.

An API client for Rainwave.cc is added to simplify walking through all
favorites. Corresponding data models are added to structure the API
results.

A downloader for OC ReMix is added to simplify downloading remixes.